feat (#33): create dashboard layout and update overview page
