--- frontend/src/pages/core/DashboardByProjectView.vue
@@ -1,9 +1,0 @@
-<template>
-    
-</template>
-
-<script>
-export default {
-    
-}
-</script>
